% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort2Trajectory.R
\name{Cohort2Trajectory_modified}
\alias{Cohort2Trajectory_modified}
\title{This function creates patient treatment trajectories}
\usage{
Cohort2Trajectory_modified(
  dbms = "postgresql",
  connection = NULL,
  cdmSchema = "ohdsi_cdm",
  cdmVocabSchema = "ohdsi_vocab",
  cdmTmpSchema = "ohdsi_temp",
  cdmResultsSchema = "ohdsi_results",
  studyName = "Cohort2Trajectory",
  baseUrl = "http://localhost:8080/WebAPI",
  stateCohortLabels = NULL,
  stateCohortPriorityOrder = NULL,
  stateCohortMandatory = NULL,
  stateCohortAbsorbing = NULL,
  stateSelectionType = NULL,
  oocFix = "None",
  trajectoryType = NULL,
  lengthOfStay = NULL,
  outOfCohortAllowed = NULL,
  runSavedStudy = FALSE,
  pathToResults = getwd(),
  useCDM = TRUE,
  pathToData = "./tmp/datasets/importedData.csv",
  allowedStatesList = createStateList(stateCohortLabels),
  mergeStates = FALSE,
  mergeThreshold = 0.5,
  runGeneration = TRUE
)
}
\arguments{
\item{dbms}{The type of DBMS running on the server. Valid values are:'oracle','postgresql','redshift','sql server','pdw','netezza','bigquery','sqlite', 'sqlite extended','spark'}

\item{connection}{Connection to database}

\item{cdmSchema}{Schema which contains the OHDSI Common Data Model.}

\item{cdmVocabSchema}{Schema which contains the OHDSI Common Data Modelvocabulary tables.}

\item{cdmTmpSchema}{Schema for temporary tables, will be deleted.}

\item{cdmResultsSchema}{Schema which has the information about thecohorts created in Atlas}

\item{stateCohortLabels}{Vector of the customized labels of the statecohorts}

\item{stateCohortPriorityOrder}{Vector of the customized labels of thestate cohorts in priority order}

\item{stateCohortMandatory}{Vector of the customized labels of thestate cohorts which are mandatory in trajectory}

\item{stateCohortAbsorbing}{Vector of the customized labels of thestate cohorts which are absorbing}

\item{stateSelectionType}{The type of state selection (1 - Firstoccurring, 2 - Max overlap, 3 - Priority)}

\item{oocFix}{The method to use for replacing "OUT OF COHORT" stateswith more relevant states}

\item{trajectoryType}{The type of the trajectory (0 - Discrete time, 1- Continuous time)}

\item{lengthOfStay}{The length of stay (days) in one state (Effect onlyin discrete case)}

\item{outOfCohortAllowed}{boolean whether the patient trajectory cansurpass the target cohort's observation-period}

\item{useCDM}{The package can also be run without the OMOP CDM}

\item{pathToData}{When using without OMOP CDM specify the path to datafile}

\item{allowedStatesList}{A list object which indicates accessiblestates from said state}

\item{mergeStates}{Boolean, if you want to merge states when theyoverlap}

\item{mergeThreshold}{Value from 0 to 1. If mergeStates is TRUE thestates will be label-merged given they overlap more than the specifiedthreshold. Can be given as vector, then multiple iterations are runned,}

\item{runGeneration}{TRUE/FALSE if FALSE no data tranfromation will berun, only the imported dataset will be cleaned}
}
\description{
This function creates patient treatment trajectories
}
