% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPhenotypeLibraryToJson.R
\name{readInclusionRules}
\alias{readInclusionRules}
\title{Read inclusion rules}
\usage{
readInclusionRules(filename, startWindowTime)
}
\arguments{
\item{filename}{(String) path to inclusion rules filename}

\item{startWindowTime}{(String) sets start window time inclusion rule value}
}
\value{
inclusion rules
}
\description{
Reads in inclusion rules file and modifies start window time of the query if needed
subfunction of readPhenotypeLibraryToJson
}
