% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPhenotypeLibraryToJson.R
\name{readPhenotypeLibraryToJson}
\alias{readPhenotypeLibraryToJson}
\title{Read phenotype library to json}
\usage{
readPhenotypeLibraryToJson(
 
    cohortsURL = "https://raw.githubusercontent.com/OHDSI/PhenotypeLibrary/main/inst/Cohorts.csv",
 
    cohortURL = "https://raw.githubusercontent.com/OHDSI/PhenotypeLibrary/main/inst/cohorts/",
  inclusionRulesDays = 365,
  setInclusionRules = TRUE,
  cohortIds = NULL,
  pathToResults
)
}
\arguments{
\item{cohortsURL}{(String) link to raw webpage containing csv with cohorts and information}

\item{cohortURL}{(String) partial link to raw webpage that contains a json file for a specific cohort}

\item{inclusionRulesDays}{(String) parameter for ATLAS JSON query StartWindow start days}

\item{setInclusionRules}{(boolean) deafault value TRUE; option to not use inclusion rules}

\item{cohortIds}{(c()) default value NULL; adding vector with cohortIds will read only those cohorts}

\item{pathToResults}{(String) path to working directory}
}
\value{

}
\description{
This functions reads x.json cohorts from OHDSI/PhenotypeLibrary github
repository. The JSON-files are adjusted to suit ATLAS queries.
}
\examples{
readPhenotypeLibraryToJson(inclusionRulesDays = "730", cohortIds = c(688,1054,785))

}
